requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
motor==3.5.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.7.4
python-dotenv==1.0.1

